Updated Wrapper to support LJM 0.0245
Finished combining readS into read.
In the process of combining writeS into write.

Added a new dictionary that allows for constant search times for
strings to reg-info for the ljm_constants.json file.

Added a "run.sh" file that runs the testMe.js file & opens sublime to
this project (on my computer) for convenience & productivity purposes.